python/ai-evaluator-py/src/ai_evaluator/config/os_environ/azure_openai.py
from pydantic import BaseModel, Field


class AzureOpenAISettings(BaseModel):
    base_url: str = Field()

    # XXX: Use AIProjectClient
    api_key: str = Field()

    deployment_name: str = Field()

    api_version: str = Field()
